requirements.txt.txt
streamlit
opencv-python-headless
numpy
ultralytics
easyocr
Pillow
